TestesComSpringApiVendas/Vendas/SpringExemplo/src/main/java/com/example/SpringExemplo/controllers/OrderController.java
package com.example.SpringExemplo.controllers;

import com.example.SpringExemplo.entites.Order;
import com.example.SpringExemplo.services.OrderService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@RestController
@RequestMapping(value = "/order")
public class OrderController {

    @Autowired
    private OrderService service;

    @GetMapping
    public ResponseEntity<List<Order>> findAll() {
        List<Order> list = service.listarPedidos();
        return ResponseEntity.ok().body(list);
    }

    @GetMapping(value = "/{id}")
    public ResponseEntity<Order> findById(@PathVariable Long id) {
        Order obj = service.listarPedidoId(id);
        return ResponseEntity.ok().body(obj);
    }



}
